requirements-dev.txt
-f requirements.txt
pytest==5.4.3
pytest-asyncio==0.12.0
pytest-aiohttp==0.3.0
pytest-mock==3.1.1
requests-mock==1.8.0
fake-rpi==0.7.1